move locale code out of seekerApplication
