+++ Seeker/SeekerApplication.cs
@@ -177,20 +177,20 @@ namespace Seeker
 
             SeekerState.OffsetFromUtcCached = SimpleHelpers.GetDateTimeNowSafe().Subtract(DateTime.UtcNow);
 
-            SeekerState.SystemLanguage = LocaleToString(Resources.Configuration.Locale);
+            SeekerState.SystemLanguage = LocaleHelper.LocaleToString(Resources.Configuration.Locale);
 
-            if (HasProperPerAppLanguageSupport())
+            if (LocaleHelper.HasProperPerAppLanguageSupport())
             {
                 if (!PreferencesState.LegacyLanguageMigrated)
                 {
                     PreferencesState.LegacyLanguageMigrated = true;
                     PreferencesManager.SaveLegacyLanguageMigrated();
-                    SetLanguage(PreferencesState.Language);
+                    LocaleHelper.SetLanguage(PreferencesState.Language);
                 }
             }
             else
             {
-                SetLanguageLegacy(PreferencesState.Language, false);
+                LocaleHelper.SetLanguageLegacy(PreferencesState.Language, false);
             }
 
             //LogDebug("Default Night Mode: " + AppCompatDelegate.DefaultNightMode); //-100 = night mode unspecified, default on my Pixel 2. also on api22 emulator it is -100.
@@ -398,147 +398,6 @@ namespace Seeker
             SearchUtil.ExcludedSearchPhrases = exludedPhrasesList;
         }
 
-        public static string GetLegacyLanguageString()
-        {
-            if (SeekerApplication.HasProperPerAppLanguageSupport())
-            {
-                var lm = (LocaleManager)Context.GetSystemService(Context.LocaleService);
-                LocaleList appLocales = lm.ApplicationLocales;
-                if (appLocales.IsEmpty)
-                {
-                    return PreferencesState.FieldLangAuto;
-                }
-                else
-                {
-                    Java.Util.Locale locale = appLocales.Get(0);
-                    string lang = locale.Language; // ex. fr, uk
-                    string country = locale.Country; // ex. BR, PT, CN
-                    if (!string.IsNullOrEmpty(country))
-                    {
-                        return lang + "-r" + country; // e.g. pt-rBR, pt-rPT, zh-rCN
-                    }
-                    return lang;
-                }
-            }
-            else
-            {
-                return PreferencesState.Language;
-            }
-        }
-
-        // TODO common.locale
-        /// <summary>
-        /// converts say "pt-rBR" to "pt-BR"
-        /// </summary>
-        /// <param name="locale"></param>
-        /// <returns></returns>
-        public static string FormatLocaleFromResourcesToStandard(string locale)
-        {
-            if (locale.Length == 6 && locale.Contains("-r"))
-            {
-                return locale.Replace("-r", "-");
-            }
-            else
-            {
-                return locale;
-            }
-        }
-
-        public static bool HasProperPerAppLanguageSupport()
-        {
-            return OperatingSystem.IsAndroidVersionAtLeast(33);
-        }
-
-        // TODO common locale
-        public static Java.Util.Locale LocaleFromString(string localeString)
-        {
-            Java.Util.Locale locale = null;
-            if (localeString.Contains("-r"))
-            {
-                var parts = localeString.Replace("-r", "-").Split('-');
-                locale = new Java.Util.Locale(parts[0], parts[1]);
-            }
-            else
-            {
-                locale = new Java.Util.Locale(localeString);
-            }
-            return locale;
-        }
-
-
-        public void SetLanguage(string language)
-        {
-            if (HasProperPerAppLanguageSupport())
-            {
-                var lm = (LocaleManager)ApplicationContext.GetSystemService(Context.LocaleService);
-
-                if (language == PreferencesState.FieldLangAuto)
-                {
-                    lm.ApplicationLocales = LocaleList.EmptyLocaleList;
-                }
-                else
-                {
-                    lm.ApplicationLocales = LocaleList.ForLanguageTags(FormatLocaleFromResourcesToStandard(language));
-                }
-            }
-            else
-            {
-                SetLanguageLegacy(PreferencesState.Language, true);
-            }
-        }
-
-        public void SetLanguageLegacy(string language, bool changed)
-        {
-            string localeString = language;
-            var res = this.Resources;
-            var config = res.Configuration;
-            var displayMetrics = res.DisplayMetrics;
-
-            var currentLocale = config.Locale;
-
-            if (LocaleToString(currentLocale) == language)
-            {
-                return;
-            }
-
-            if (language == PreferencesState.FieldLangAuto && SeekerState.SystemLanguage == LocaleToString(currentLocale))
-            {
-                return;
-            }
-
-
-            Java.Util.Locale locale = language != PreferencesState.FieldLangAuto ? LocaleFromString(localeString) : LocaleFromString(SeekerState.SystemLanguage);
-
-            Java.Util.Locale.Default = locale;
-            config.SetLocale(locale);
-
-            this.BaseContext.Resources.UpdateConfiguration(config, displayMetrics);
-
-            if (changed)
-            {
-                RecreateActivies();
-            }
-        }
-
-        // TODO Common.LocaleHelper
-        public static string LocaleToString(Java.Util.Locale locale)
-        {
-            //"en" ""
-            //"pt" "br"
-            if (string.IsNullOrEmpty(locale.Variant))
-            {
-                return locale.Language;
-            }
-            else
-            {
-                return locale.Language + "-r" + locale.Variant.ToUpper();
-            }
-        }
-
-        public static bool AreLocalesSame(Java.Util.Locale locale1, Java.Util.Locale locale2)
-        {
-            return LocaleToString(locale1) == LocaleToString(locale2);
-        }
 
         /// <summary>
         ///
